Refactor core
