+++ src/core/mod.rs
@@ -2,3 +2,8 @@ pub mod args;
 pub mod idf;
 pub mod min_max_tie_heap;
 pub mod name;
+
+pub use args::*;
+pub use idf::*;
+pub use min_max_tie_heap::*;
+pub use name::*;
